+++ pkg/utils/stdliblogtolog/adapter_test.go
@@ -1,4 +1,4 @@
-package stdliblogtologrus
+package stdliblogtolog
 
 import (
 	"bufio"
